Updated sample output, added check for invalid input, formatted code, and moved to correct package.
